% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCR.R
\name{JMAn}
\alias{JMAn}
\title{A function computing the  Jackknife Model Average (JMA) least-squares estimates for pure nested candimate models.}
\usage{
JMAn(y, x)
}
\arguments{
\item{y}{n×1 dependent varables}

\item{x}{n×p regressor matrix}
}
\value{
A list inluding parameter estimate, weight vector,fitted values,fitted residuals and Value of  Cross-Validation criterion.
}
\description{
A function computing the Jackknife Model Average (JMA) least-squares estimates for pure nested candimate models.
}
\examples{
\dontrun{
y<-matrix(runif(10),nrow=10,ncol=1)
x<-matrix(rnorm(70),nrow=10,ncol=7)
MMAn(y,x)
}
}
