% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwc}
\alias{rwc}
\title{A Metropolis sampler using Rcpp}
\usage{
rwc(sigma, x, N)
}
\arguments{
\item{sigma}{variance}

\item{x}{the number of between-sample random number}

\item{N}{the number of samples}
}
\value{
a numericmatrix with a random sample of size \code{n} and k
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
sigma <- 2
x <- 25
N <- 2000
res <- rwc(sigma, x, N)
}
}
