\name{StatComp20034-package}
\alias{StatComp20034-package}
\alias{StatComp20034}
\docType{package}
\title{
   Statistics Computation Course Final Project
}
\description{
  Statistics Computation Course Final Project
  Include: two functions about model averaging and all homeworks with rmd and html.
}
\details{
  The Mallows Model Averaging least-squares estimates is descripted by " Hansen, B.E., 2007. Least squares model averaging. Econometrica 75."
  The Jackknife Model Averaging least-squares estimates is descripted by"Hansen, B.E., Racine, J., 2012. Jackknife model averaging. Journal of Econometrics 167."
}
\author{
Luyao Wang

Maintainer: Luyao Wang <wly5531@mail.ustc.edu.cn>
}
\references{

1.Hansen, B.E., 2007. Least squares model averaging. Econometrica 75

2.Hansen, B.E., Racine, J., 2012. Jackknife model averaging. Journal of Econometrics 167
}
\keyword{ package }
